+++ azure-toolkit-libs/azure-toolkit-auth-lib/src/main/java/com/microsoft/azure/tools/auth/core/ManagedIdentityCredentialRetriever.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.azure.tools.auth.tools;
+package com.microsoft.azure.tools.auth.core;
 
 import com.azure.core.credential.TokenRequestContext;
 import com.azure.identity.CredentialUnavailableException;
